+++ components/VideoModal.tsx
@@ -1,5 +1,10 @@
 import React from 'react';
-import { Modal } from 'react-bootstrap';
+import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle
+} from '@/components/ui/dialog';
 
 interface VideoModalProps {
   show: boolean;
@@ -9,11 +14,11 @@ interface VideoModalProps {
 
 export const VideoModal: React.FC<VideoModalProps> = ({ show, onHide, videoId }) => {
   return (
-    <Modal show={show} onHide={onHide} size="lg" centered>
-      <Modal.Header closeButton>
-        <Modal.Title>Video</Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
+    <Dialog open={show} onOpenChange={(open) => !open && onHide()}>
+      <DialogContent className="sm:max-w-3xl">
+        <DialogHeader>
+          <DialogTitle>Video</DialogTitle>
+        </DialogHeader>
         {videoId && (
           <iframe
             width="100%"
@@ -25,7 +30,7 @@ export const VideoModal: React.FC<VideoModalProps> = ({ show, onHide, videoId })
             title="Embedded Video"
           ></iframe>
         )}
-      </Modal.Body>
-    </Modal>
+      </DialogContent>
+    </Dialog>
   );
 };