+++ pages/_app.tsx
@@ -1,12 +1,17 @@
 import '@/styles/globals.css'
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
+import { ThemeProvider } from '@/components/ThemeProvider'
+
 export default function App({ Component, pageProps }: AppProps) {
   return (
-  <>
-  <Head>
-    <meta name="viewport" content="width=device-width, initial-scale=1"></meta>
-  </Head>
-    <Component {...pageProps} /> 
-  </>)
+    <>
+      <Head>
+        <meta name="viewport" content="width=device-width, initial-scale=1"></meta>
+      </Head>
+      <ThemeProvider>
+        <Component {...pageProps} />
+      </ThemeProvider>
+    </>
+  )
 }